Add default environment file (.envdefaults).
dotenv-hs doesn't support shell interpolation yet, but you can have two
environment files, the second of which can specify any variables in the
first which use shell interpolation (which ARKSINK_DB_USERNAME does).

Also, variables set in the first file and not in the second don't exist
in the environment set by dotenv-hs. I am still committing this file to
the project in anticipation of the day that it all just works (or I use
something sane like the dotenv Ruby gem).