Fix inability to sign in again after signing out.
This occurred because, when signing in again, a new random IV would be
used to encrypt the cookie, making the name of the cookie different to
the previous one.

This would leave > 1 cookie behind, causing conflict after decryption.
Scotty.Cookie would return the value of the first one, which turns out
to be the one set earlier, i.e. from the first login.

This can be fixed by using a null IV when encrypting; this returns the
same ciphertext for the same plaintext encrypted with the same key.