Remove .envdefaults dotenv file.
There are various reasons for this change.

This file was initially created to serve as both sensible defaults for
variables which have some, so that users can get up and running quick,
and documentation on what variables need to be provided.

dotenv-hs now provides a flag, `--example`, which takes a file listing
all the required keys (without values) and kills the process if any of
them are absent, resulting in more sensical errors.

The default zeroed validation & encryption key values are not provided
by default any more, which preclude the app running in production with
insecure keys by accident.

In 1b967c1 (the commit introducing this file), I claim, “variables set
in the first file and not in the second don't exist in the environment
set by dotenv-hs”; it seems false from some testing I did, but I could
be misunderstanding it.

The trigger for this change was that I wanted to override variables in
the environment from `.env`, but not `.envdefaults`; it is supposed to
work that way anyway, but it only bit me when I tried to Dockerise the
app and the `.envdefaults` values was overriding the DB config values,
which were being set in `docker-compose.yml`, most notably the host of
the database container being reset to localhost.